Search many locations for unqualified dot program
If the name of the dot executable is not qualified with a full path, we
should search through some likely locations for it. Use the 'env'
program with some likely paths to search through.

(A better way to know where to search would be to get the path from the
user's environment, but the javascript runtime seems to scour that away.)

If we fail to complete running, tell the user what the problem probably is.
From env man page,
`An exit status of 127 indicates that utility could not be found.`

Closes: #7
Closes: #20